requirements.txt
# HTTP requests
requests>=2.31.0

# Scheduling
schedule>=1.2.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Type hints
typing-extensions>=4.8.0

# Logging
python-json-logger>=2.0.7

# Environment variables
python-dotenv>=1.0.0
